Move the keyDown replay guard into its own file to satisfy the Swift file length budget
No behavior change. The guard helper, identity struct, and in-flight set
move from Sources/AppDelegate.swift (which the fix had pushed 32 lines
over its 18057-line budget) into Sources/App/WindowKeyDownReplayGuard.swift,
leaving AppDelegate.swift 35 lines under budget. The new file stays below
the 500-line tracking threshold.
