+++ Sources/AppDelegate.swift
@@ -16552,36 +16552,6 @@ private var cmuxFirstResponderGuardHitViewOverride: NSView?
 private var cmuxFirstResponderGuardCurrentEventContext: NSEvent?
 private var cmuxFirstResponderGuardHitViewContext: NSView?
 private var cmuxFirstResponderGuardContextWindowNumber: Int?
-/// Identity of a key event currently being force-dispatched into a responder's
-/// `keyDown(with:)` by `NSWindow.cmux_performKeyEquivalent(with:)`.
-///
-/// Forwarding keyDown can re-enter `performKeyEquivalent` with the same event
-/// while the dispatch is still on the stack: WebKit replays unhandled keys
-/// through the responder chain, and on macOS 26 `-[NSWindow keyDown:]`
-/// re-enters `performKeyEquivalent`. Without a replay guard at the dispatch
-/// chokepoint the same event ping-pongs between the swizzle and the focused
-/// responder until the main-thread stack overflows
-/// (https://github.com/manaflow-ai/cmux/issues/5887).
-///
-/// Identity is the event's stable field tuple rather than object identity so
-/// the guard still holds if AppKit/WebKit re-deliver the event as an equal
-/// copy. Key autorepeat produces distinct events (fresh timestamps), so
-/// repeat typing is never throttled. The dispatching window's number is part
-/// of the identity so windows cannot suppress each other's dispatches.
-private struct CmuxForceDispatchedKeyEventIdentity: Hashable {
-    let windowNumber: Int
-    let eventType: UInt
-    let keyCode: UInt16
-    let modifierFlags: UInt
-    let timestamp: TimeInterval
-}
-
-/// Events whose force-dispatch is currently on the main-thread stack.
-/// Main-thread only (key-event dispatch); entries are stack-scoped, inserted
-/// before `keyDown(with:)` and removed when the dispatch unwinds, so WebKit's
-/// legitimate replay of an unhandled key (which arrives after the original
-/// dispatch has fully unwound) is still force-dispatched normally.
-private var cmuxInFlightForceDispatchedKeyEventIdentities = Set<CmuxForceDispatchedKeyEventIdentity>()
 private var cmuxWindowFirstResponderBypassDepth = 0
 private var cmuxFieldEditorOwningWebViewAssociationKey: UInt8 = 0
 
@@ -17214,43 +17184,6 @@ private extension NSWindow {
             originalDispatchMs = (ProcessInfo.processInfo.systemUptime - originalDispatchStart) * 1000.0
         }
 #endif
-    }
-
-    /// Single chokepoint for every direct `keyDown(with:)` force-dispatch made
-    /// by `cmux_performKeyEquivalent(with:)`.
-    ///
-    /// Dispatches `event` into `target`'s `keyDown(with:)` unless the same
-    /// event is already being force-dispatched lower on this window's call
-    /// stack, and returns whether the dispatch happened. Callers that get
-    /// `false` back must decline the event (fall through to default AppKit
-    /// handling) instead of dispatching themselves; re-dispatching the same
-    /// in-flight event is the infinite key-routing loop from
-    /// https://github.com/manaflow-ai/cmux/issues/5887.
-    private func cmuxForceDispatchKeyDownOnce(
-        _ event: NSEvent,
-        to target: NSResponder,
-        reason: @autoclosure () -> String
-    ) -> Bool {
-        let identity = CmuxForceDispatchedKeyEventIdentity(
-            windowNumber: self.windowNumber,
-            eventType: event.type.rawValue,
-            keyCode: event.keyCode,
-            modifierFlags: event.modifierFlags.rawValue,
-            timestamp: event.timestamp
-        )
-        guard !cmuxInFlightForceDispatchedKeyEventIdentities.contains(identity) else {
-#if DEBUG
-            cmuxDebugLog("  → \(reason()) reentry; declining force-dispatch of in-flight key event")
-#endif
-            return false
-        }
-        cmuxInFlightForceDispatchedKeyEventIdentities.insert(identity)
-        defer { cmuxInFlightForceDispatchedKeyEventIdentities.remove(identity) }
-#if DEBUG
-        cmuxDebugLog("  → \(reason()) routed to firstResponder.keyDown")
-#endif
-        target.keyDown(with: event)
-        return true
     }
 
     @objc func cmux_performKeyEquivalent(with event: NSEvent) -> Bool {
